Add active links
