+++ src/components/layout/Header.tsx
@@ -1,5 +1,4 @@
-import Link from "next/link";
-import React from "react";
+import Navigation from "../ui/Navigation";
 
 export default function Header() {
   return (
@@ -10,24 +9,8 @@ export default function Header() {
           / FIND YOUR LAST BARBER
         </div>
       </div>
-      <ul className="flex gap-4">
-        {LINKS.map((link) => (
-          <li
-            key={link.id}
-            className="uppercase text-sm text-white/40 hover:text-white"
-          >
-            <Link href={link.href}>/ {link.text}</Link>
-          </li>
-        ))}
-      </ul>
+     <Navigation />
     </div>
   );
 }
 
-const LINKS = [
-  { id: 1, text: "home", href: "/" },
-  { id: 2, text: "barbers", href: "/barbers" },
-  { id: 3, text: "join", href: "/join" },
-  { id: 4, text: "login", href: "/login" },
-  { id: 5, text: "about us", href: "/about us" },
-];
